pyproject.toml
[tool.poetry]
name = "pulse"
version = "0.1.0"
description = ""
authors = ["Fraser <fraserisbester@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.338"
faiss-cpu = "^1.7.4"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.5.1"
langchainhub = "^0.1.14"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
python-dotenv = "^1.0.0"
slack-sdk = "^3.24.0"
python-multipart = "^0.0.6"
pydantic = "^2.5.1"
openai = "^1.3.3"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
